requirements.txt
# Telegram Bot
aiogram==3.4.1

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqladmin[fastapi]==0.16.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Admin Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
itsdangerous==2.1.2

# Security & Rate Limiting
slowapi==0.1.9

# Structured Logging (для Phase 3)
structlog==23.2.0

# Testing (для Phase 4)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.26.0
